Split worker lib by concern: dice authority + auth provisioning modules
- lib/dice.ts: move authoritativeRollDice out of localDev (it's a production
  security boundary, not a dev-only helper); localDev.ts is now purely the
  loopback request gate.
- lib/authProvisioning.ts: move generateId/generateToken/SESSION_TTL_MS +
  provisionNewUserStatements + issueAuthSession out of the route file; auth.ts
  drops from 611 to 533 lines and stays focused on handlers/WebAuthn.
- update imports (games.ts, auth.ts, localDev unit test).
Validated: ci:local green (411 unit + 154 integration), live dev-login + WS smoke.
