+++ packages/worker/src/routes/auth.ts
@@ -1,7 +1,6 @@
 import { zValidator } from "@hono/zod-validator";
 import {
   AuthErrorKeys,
-  type AuthSessionResponse,
   DevLoginInputSchema,
   LoginOptionsInputSchema,
   LoginVerifyInputSchema,
@@ -22,6 +21,11 @@ import {
   verifyRegistrationResponse,
 } from "@simplewebauthn/server";
 import { Hono } from "hono";
+import {
+  generateId,
+  issueAuthSession,
+  provisionNewUserStatements,
+} from "../lib/authProvisioning.js";
 import { isLocalDevRequest } from "../lib/localDev.js";
 
 type Bindings = {
@@ -57,89 +61,7 @@ function getExpectedOrigin(c: { env?: Bindings }): string {
   return c.env?.WEBAUTHN_ORIGIN ?? "http://localhost:5173";
 }
 
-function generateId(): string {
-  const bytes = new Uint8Array(16);
-  crypto.getRandomValues(bytes);
-  return Array.from(bytes)
-    .map((b) => b.toString(16).padStart(2, "0"))
-    .join("");
-}
-
-function generateToken(): string {
-  const bytes = new Uint8Array(32);
-  crypto.getRandomValues(bytes);
-  return Array.from(bytes)
-    .map((b) => b.toString(16).padStart(2, "0"))
-    .join("");
-}
-
-const SESSION_TTL_MS = 30 * 24 * 60 * 60 * 1000; // 30 days
 const CHALLENGE_TTL_SECONDS = 300; // 5 minutes
-
-/**
- * Prepared statements that bootstrap a new user plus its companion rows
- * (visibility / rank / trustworthiness). Returned (not executed) so callers can
- * include them in their own atomic `db.batch(...)` alongside path-specific rows
- * (e.g. a passkey credential for registration, or just a session for
- * dev-login). Single source of truth for the "new user" shape so register and
- * dev-login stay consistent when the schema changes.
- */
-function provisionNewUserStatements(
-  db: D1Database,
-  userId: string,
-  username: string,
-  now: number,
-): D1PreparedStatement[] {
-  return [
-    db
-      .prepare(
-        "INSERT INTO users (id, username, locale, theme_preference, created_at, updated_at, role) VALUES (?, ?, 'en', 'system', ?, ?, 'user')",
-      )
-      .bind(userId, username, now, now),
-    db.prepare("INSERT INTO user_visibility (user_id) VALUES (?)").bind(userId),
-    db
-      .prepare(
-        "INSERT INTO user_ranks (user_id, tier, rank_points) VALUES (?, 0, 0)",
-      )
-      .bind(userId),
-    db
-      .prepare(
-        "INSERT INTO trustworthiness (user_id, score, last_updated_at) VALUES (?, 7, ?)",
-      )
-      .bind(userId, now),
-  ];
-}
-
-/**
- * Mint a session for a user: the `auth_sessions` INSERT plus expired-session
- * cleanup, returned as prepared statements so callers batch them atomically
- * alongside any path-specific rows, together with the standard session response
- * payload. Single source of truth so register / login / dev-login can't drift.
- */
-function issueAuthSession(
-  db: D1Database,
-  userId: string,
-  username: string,
-  now: number,
-): { statements: D1PreparedStatement[]; response: AuthSessionResponse } {
-  const token = generateToken();
-  const expiresAt = now + SESSION_TTL_MS;
-  return {
-    statements: [
-      db
-        .prepare(
-          "INSERT INTO auth_sessions (id, user_id, token, expires_at, created_at) VALUES (?, ?, ?, ?, ?)",
-        )
-        .bind(generateId(), userId, token, expiresAt, now),
-      db
-        .prepare(
-          "DELETE FROM auth_sessions WHERE user_id = ? AND expires_at < ?",
-        )
-        .bind(userId, now),
-    ],
-    response: { token, userId, username, expiresAt },
-  };
-}
 
 interface CredentialRow {
   id: string;
